Remove ngcontrol ref to startScore.
It's not really needed since you are already inside the RefCtrl
controller. You don't really want to nest the ideas of controller.

This fixes the text fields from not working correctly when you change
the score manually and then try to use the buttons. I believe the use of
ngcontrol on something that isn't really a controller screws up some of
the $scope bindings.

I left the text fields disabled since they don't really make sense to be
directly modifiable for a mobile targeted app.